lets loops run multithreaded
