+++ include/lingodb/compiler/Dialect/SubOperator/Transforms/Passes.h
@@ -14,9 +14,9 @@ std::unique_ptr<mlir::Pass> createReuseLocalPass();
 std::unique_ptr<mlir::Pass> createParallelizePass();
 std::unique_ptr<mlir::Pass> createSpecializeParallelPass();
 std::unique_ptr<mlir::Pass> createSplitIntoExecutionStepsPass();
+std::unique_ptr<mlir::Pass> createPrepareLoweringPass(bool loopsOnly = false);
 std::unique_ptr<mlir::Pass> createInlineNestedMapPass();
 std::unique_ptr<mlir::Pass> createFinalizePass();
-std::unique_ptr<mlir::Pass> createPrepareLoweringPass();
 std::unique_ptr<mlir::Pass> createCommonPiplineEliminationPass();
 std::unique_ptr<mlir::Pass> createHoistInvariantStatePass();
 void registerSubOpTransformations();
