+++ src/compiler/Dialect/SubOperator/Transforms/Passes.cpp
@@ -36,6 +36,9 @@ void subop::registerSubOpTransformations() {
       return subop::createSplitIntoExecutionStepsPass();
    });
    ::mlir::registerPass([]() -> std::unique_ptr<::mlir::Pass> {
+      return subop::createPrepareLoweringPass();
+   });
+   ::mlir::registerPass([]() -> std::unique_ptr<::mlir::Pass> {
       return subop::createCommonPiplineEliminationPass();
    });
    ::mlir::registerPass([]() -> std::unique_ptr<::mlir::Pass> {
